+++ docker/test_env/nethermind.go
@@ -1,83 +1,413 @@
 package test_env
 
 import (
+	"context"
+	"encoding/hex"
 	"fmt"
+	"os"
+	"strings"
 	"testing"
+	"time"
 
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/stretchr/testify/require"
+	"github.com/docker/docker/api/types/container"
+	"github.com/docker/docker/api/types/mount"
+	"github.com/google/uuid"
+	"github.com/rs/zerolog"
+
+	tc "github.com/testcontainers/testcontainers-go"
+	tcwait "github.com/testcontainers/testcontainers-go/wait"
 
 	"github.com/smartcontractkit/chainlink-testing-framework/blockchain"
+	"github.com/smartcontractkit/chainlink-testing-framework/docker"
 	"github.com/smartcontractkit/chainlink-testing-framework/logging"
-	"github.com/smartcontractkit/chainlink-testing-framework/utils/testcontext"
+	"github.com/smartcontractkit/chainlink-testing-framework/mirror"
+	"github.com/smartcontractkit/chainlink-testing-framework/utils/templates"
 )
 
-func TestEth1WithNethermind(t *testing.T) {
-	l := logging.GetTestLogger(t)
+const defaultNethermindPosImage = "nethermind/nethermind:1.25.1"
+const defaultNethermindPoaImage = "nethermind/nethermind:1.16.0"
 
-	builder := NewEthereumNetworkBuilder()
-	cfg, err := builder.
-		WithConsensusType(ConsensusType_PoW).
-		WithExecutionLayer(ExecutionLayer_Nethermind).
-		Build()
-	require.NoError(t, err, "Builder validation failed")
-
-	net, _, err := cfg.Start()
-	require.NoError(t, err, "Couldn't start PoW network")
-
-	c, err := blockchain.ConnectEVMClient(net, l)
-	require.NoError(t, err, "Couldn't connect to the evm client")
-
-	address := common.HexToAddress("0x90F8bf6A479f320ead074411a4B0e7944Ea8c9C1")
-	err = sendAndCompareBalances(testcontext.Get(t), c, address)
-	require.NoError(t, err, fmt.Sprintf("balance wasn't correctly updated when %s network", net.Name))
-
-	err = c.Close()
-	require.NoError(t, err, "Couldn't close the client")
+type Nethermind struct {
+	EnvComponent
+	ExternalHttpUrl      string
+	InternalHttpUrl      string
+	ExternalWsUrl        string
+	InternalWsUrl        string
+	InternalExecutionURL string
+	ExternalExecutionURL string
+	generatedDataHostDir string
+	chainConfg           *EthereumChainConfig
+	consensusLayer       ConsensusLayer
+	l                    zerolog.Logger
+	t                    *testing.T
 }
 
-func TestEth2WithPrysmAndNethermind(t *testing.T) {
-	l := logging.GetTestLogger(t)
+func NewNethermindPos(networks []string, generatedDataHostDir string, consensusLayer ConsensusLayer, opts ...EnvComponentOption) (*Nethermind, error) {
+	parts := strings.Split(defaultNethermindPosImage, ":")
+	g := &Nethermind{
+		EnvComponent: EnvComponent{
+			ContainerName:    fmt.Sprintf("%s-%s", "nethermind-pos", uuid.NewString()[0:8]),
+			Networks:         networks,
+			ContainerImage:   parts[0],
+			ContainerVersion: parts[1],
+		},
+		generatedDataHostDir: generatedDataHostDir,
+		consensusLayer:       consensusLayer,
+		l:                    logging.GetTestLogger(nil),
+	}
+	for _, opt := range opts {
+		opt(&g.EnvComponent)
+	}
+	// if the internal docker repo is set then add it to the version
+	g.EnvComponent.ContainerImage = mirror.AddMirrorToImageIfSet(g.EnvComponent.ContainerImage)
+	return g, nil
+}
 
-	builder := NewEthereumNetworkBuilder()
-	cfg, err := builder.
-		WithConsensusType(ConsensusType_PoS).
-		WithConsensusLayer(ConsensusLayer_Prysm).
-		WithExecutionLayer(ExecutionLayer_Nethermind).
-		Build()
-	require.NoError(t, err, "Builder validation failed")
+func NewNethermindPoa(networks []string, chainConfg *EthereumChainConfig, opts ...EnvComponentOption) (*Nethermind, error) {
+	parts := strings.Split(defaultNethermindPoaImage, ":")
+	g := &Nethermind{
+		EnvComponent: EnvComponent{
+			ContainerName:    fmt.Sprintf("%s-%s", "nethermind-poa", uuid.NewString()[0:8]),
+			Networks:         networks,
+			ContainerImage:   parts[0],
+			ContainerVersion: parts[1],
+		},
+		chainConfg: chainConfg,
+		l:          logging.GetTestLogger(nil),
+	}
+	for _, opt := range opts {
+		opt(&g.EnvComponent)
+	}
+	// if the internal docker repo is set then add it to the version
+	g.EnvComponent.ContainerImage = mirror.AddMirrorToImageIfSet(g.EnvComponent.ContainerImage)
+	return g, nil
+}
 
-	_, eth2, err := cfg.Start()
-	require.NoError(t, err, "Couldn't start PoS network")
+func (g *Nethermind) WithTestInstance(t *testing.T) ExecutionClient {
+	g.l = logging.GetTestLogger(t)
+	g.t = t
+	return g
+}
 
-	nonEip1559Network := blockchain.SimulatedEVMNetwork
-	nonEip1559Network.Name = "Simulated Nethermind + Prysm (non-EIP 1559)"
-	nonEip1559Network.URLs = eth2.PublicWsUrls()
-	clientOne, err := blockchain.ConnectEVMClient(nonEip1559Network, l)
-	require.NoError(t, err, "Couldn't connect to the evm client")
+func (g *Nethermind) StartContainer() (blockchain.EVMNetwork, error) {
+	var r *tc.ContainerRequest
+	var err error
+	if g.GetEthereumVersion() == EthereumVersion_Eth1 {
+		r, err = g.getPowContainerRequest()
 
-	t.Cleanup(func() {
-		err = clientOne.Close()
-		require.NoError(t, err, "Couldn't close the client")
+	} else {
+		r, err = g.getPosContainerRequest()
+	}
+	if err != nil {
+		return blockchain.EVMNetwork{}, err
+	}
+
+	l := logging.GetTestContainersGoTestLogger(g.t)
+	ct, err := docker.StartContainerWithRetry(g.l, tc.GenericContainerRequest{
+		ContainerRequest: *r,
+		Reuse:            true,
+		Started:          true,
+		Logger:           l,
 	})
+	if err != nil {
+		return blockchain.EVMNetwork{}, fmt.Errorf("cannot start nethermind container: %w", err)
+	}
 
-	ctx := testcontext.Get(t)
-	address := common.HexToAddress("0x90F8bf6A479f320ead074411a4B0e7944Ea8c9C1")
-	err = sendAndCompareBalances(ctx, clientOne, address)
-	require.NoError(t, err, fmt.Sprintf("balance wasn't correctly updated when %s network", nonEip1559Network.Name))
+	host, err := GetHost(context.Background(), ct)
+	if err != nil {
+		return blockchain.EVMNetwork{}, err
+	}
+	if err != nil {
+		return blockchain.EVMNetwork{}, err
+	}
+	httpPort, err := ct.MappedPort(context.Background(), NatPort(DEFAULT_EVM_NODE_HTTP_PORT))
+	if err != nil {
+		return blockchain.EVMNetwork{}, err
+	}
+	wsPort, err := ct.MappedPort(context.Background(), NatPort(DEFAULT_EVM_NODE_WS_PORT))
+	if err != nil {
+		return blockchain.EVMNetwork{}, err
+	}
 
-	eip1559Network := blockchain.SimulatedEVMNetwork
-	eip1559Network.Name = "Simulated Nethermind + Prysm (EIP 1559)"
-	eip1559Network.SupportsEIP1559 = true
-	eip1559Network.URLs = eth2.PublicWsUrls()
-	clientTwo, err := blockchain.ConnectEVMClient(eip1559Network, l)
-	require.NoError(t, err, "Couldn't connect to the evm client")
+	if g.GetEthereumVersion() == EthereumVersion_Eth2 {
+		executionPort, err := ct.MappedPort(context.Background(), NatPort(ETH2_EXECUTION_PORT))
+		if err != nil {
+			return blockchain.EVMNetwork{}, err
+		}
+		g.InternalExecutionURL = FormatHttpUrl(g.ContainerName, ETH2_EXECUTION_PORT)
+		g.ExternalExecutionURL = FormatHttpUrl(host, executionPort.Port())
+	}
 
-	t.Cleanup(func() {
-		err = clientTwo.Close()
-		require.NoError(t, err, "Couldn't close the client")
-	})
+	g.Container = ct
+	g.ExternalHttpUrl = FormatHttpUrl(host, httpPort.Port())
+	g.InternalHttpUrl = FormatHttpUrl(g.ContainerName, DEFAULT_EVM_NODE_HTTP_PORT)
+	g.ExternalWsUrl = FormatWsUrl(host, wsPort.Port())
+	g.InternalWsUrl = FormatWsUrl(g.ContainerName, DEFAULT_EVM_NODE_WS_PORT)
 
-	err = sendAndCompareBalances(ctx, clientTwo, address)
-	require.NoError(t, err, fmt.Sprintf("balance wasn't correctly updated when %s network", eip1559Network.Name))
+	networkConfig := blockchain.SimulatedEVMNetwork
+	if g.GetEthereumVersion() == EthereumVersion_Eth1 {
+		networkConfig.Name = "Simulated Eth-1-PoA (nethermind)"
+		networkConfig.GasEstimationBuffer = 100_000_000_000
+	} else {
+		networkConfig.Name = fmt.Sprintf("Simulated Eth-2-PoS (nethermind + %s)", g.consensusLayer)
+	}
+	networkConfig.URLs = []string{g.ExternalWsUrl}
+	networkConfig.HTTPURLs = []string{g.ExternalHttpUrl}
+
+	g.l.Info().Str("containerName", g.ContainerName).
+		Msg("Started Nethermind container")
+
+	return networkConfig, nil
+}
+
+func (g *Nethermind) getPowContainerRequest() (*tc.ContainerRequest, error) {
+	keystoreDir, err := os.MkdirTemp("", "keystore")
+	if err != nil {
+		return nil, err
+	}
+
+	generatedData, err := generateKeystoreAndExtraData(keystoreDir)
+	if err != nil {
+		return nil, err
+	}
+
+	genesisJsonStr, err := templates.NethermindPoAGenesisJsonTemplate{
+		ChainId:     fmt.Sprintf("%d", g.chainConfg.ChainID),
+		AccountAddr: RootFundingAddr,
+		ExtraData:   fmt.Sprintf("0x%s", hex.EncodeToString(generatedData.extraData)),
+	}.String()
+	if err != nil {
+		return nil, err
+	}
+	genesisFile, err := os.CreateTemp("", "genesis_json")
+	if err != nil {
+		return nil, err
+	}
+	_, err = genesisFile.WriteString(genesisJsonStr)
+	if err != nil {
+		return nil, err
+	}
+
+	// create empty cfg file since if we don't pass any
+	// default mainnet.cfg will be used
+	noneCfg, err := os.CreateTemp("", "none.cfg")
+	if err != nil {
+		return nil, err
+	}
+
+	_, err = noneCfg.WriteString("{}")
+	if err != nil {
+		return nil, err
+	}
+
+	passFile, err := os.CreateTemp("", "password.txt")
+	if err != nil {
+		return nil, err
+	}
+
+	err = os.WriteFile(passFile.Name(), []byte(""), 0600)
+	if err != nil {
+		return nil, err
+	}
+
+	rootFile, err := os.CreateTemp(keystoreDir, RootFundingAddr)
+	if err != nil {
+		return nil, err
+	}
+	_, err = rootFile.WriteString(RootFundingWallet)
+	if err != nil {
+		return nil, err
+	}
+
+	return &tc.ContainerRequest{
+		Name:            g.ContainerName,
+		Image:           g.GetImageWithVersion(),
+		Networks:        g.Networks,
+		AlwaysPullImage: true,
+		// ImagePlatform: "linux/x86_64",  //don't even try this on Apple Silicon, the node won't start due to .NET error
+		ExposedPorts: []string{NatPortFormat(DEFAULT_EVM_NODE_HTTP_PORT), NatPortFormat(DEFAULT_EVM_NODE_WS_PORT)},
+		WaitingFor: tcwait.ForAll(
+			tcwait.ForLog("Nethermind initialization completed").
+				WithStartupTimeout(120 * time.Second).
+				WithPollInterval(1 * time.Second),
+		),
+		Cmd: []string{
+			"--config=/none.cfg",
+			"--Init.ChainSpecPath=/chainspec.json",
+			"--Init.DiscoveryEnabled=false",
+			"--Init.WebSocketsEnabled=true",
+			fmt.Sprintf("--JsonRpc.WebSocketsPort=%s", DEFAULT_EVM_NODE_WS_PORT),
+			"--JsonRpc.Enabled=true",
+			"--JsonRpc.EnabledModules=net,consensus,eth,subscribe,web3,admin,trace,txpool",
+			"--JsonRpc.Host=0.0.0.0",
+			fmt.Sprintf("--JsonRpc.Port=%s", DEFAULT_EVM_NODE_HTTP_PORT),
+			"--KeyStore.KeyStoreDirectory=/keystore",
+			fmt.Sprintf("--KeyStore.BlockAuthorAccount=%s", generatedData.minerAccount.Address.Hex()),
+			fmt.Sprintf("--KeyStore.UnlockAccounts=%s", generatedData.minerAccount.Address.Hex()),
+			"--KeyStore.PasswordFiles=/password.txt",
+			"--Mining.Enabled=true",
+			"--Init.PeerManagerEnabled=false",
+			"--HealthChecks.Enabled=true", // default slug /health
+		},
+		Files: []tc.ContainerFile{
+			{
+				HostFilePath:      genesisFile.Name(),
+				ContainerFilePath: "/chainspec.json",
+				FileMode:          0644,
+			},
+			{
+				HostFilePath:      noneCfg.Name(),
+				ContainerFilePath: "/none.cfg",
+				FileMode:          0644,
+			},
+			{
+				HostFilePath:      passFile.Name(),
+				ContainerFilePath: "/password.txt",
+				FileMode:          0644,
+			},
+		},
+		HostConfigModifier: func(hostConfig *container.HostConfig) {
+			hostConfig.Mounts = append(hostConfig.Mounts, mount.Mount{
+				Type:     mount.TypeBind,
+				Source:   keystoreDir,
+				Target:   "/keystore",
+				ReadOnly: false,
+			})
+		},
+		LifecycleHooks: []tc.ContainerLifecycleHooks{
+			{
+				PostStarts: g.PostStartsHooks,
+				PostStops:  g.PostStopsHooks,
+			},
+		},
+	}, nil
+}
+
+func (g *Nethermind) GetInternalExecutionURL() string {
+	if g.GetEthereumVersion() == EthereumVersion_Eth1 {
+		panic("eth1 node doesn't have an execution URL")
+	}
+	return g.InternalExecutionURL
+}
+
+func (g *Nethermind) GetExternalExecutionURL() string {
+	if g.GetEthereumVersion() == EthereumVersion_Eth1 {
+		panic("eth1 node doesn't have an execution URL")
+	}
+	return g.ExternalExecutionURL
+}
+
+func (g *Nethermind) GetInternalHttpUrl() string {
+	return g.InternalHttpUrl
+}
+
+func (g *Nethermind) GetInternalWsUrl() string {
+	return g.InternalWsUrl
+}
+
+func (g *Nethermind) GetExternalHttpUrl() string {
+	return g.ExternalHttpUrl
+}
+
+func (g *Nethermind) GetExternalWsUrl() string {
+	return g.ExternalWsUrl
+}
+
+func (g *Nethermind) GetContainerName() string {
+	return g.ContainerName
+}
+
+func (g *Nethermind) GetContainer() *tc.Container {
+	return &g.Container
+}
+
+func (g *Nethermind) GetEthereumVersion() EthereumVersion {
+	if g.consensusLayer != "" {
+		return EthereumVersion_Eth2
+	}
+
+	return EthereumVersion_Eth1
+}
+
+func (g *Nethermind) WaitUntilChainIsReady(ctx context.Context, waitTime time.Duration) error {
+	if g.GetEthereumVersion() == EthereumVersion_Eth1 {
+		return nil
+	}
+	waitForFirstBlock := tcwait.NewLogStrategy("Improved post-merge block").WithPollInterval(1 * time.Second).WithStartupTimeout(waitTime)
+	return waitForFirstBlock.WaitUntilReady(ctx, *g.GetContainer())
+}
+
+func (g *Nethermind) GetContainerType() ContainerType {
+	return ContainerType_Nethermind
+}
+
+func (g *Nethermind) getPosContainerRequest() (*tc.ContainerRequest, error) {
+	// create empty cfg file since if we don't pass any
+	// default mainnet.cfg will be used
+	noneCfg, err := os.CreateTemp("", "none.cfg")
+	if err != nil {
+		return nil, err
+	}
+
+	_, err = noneCfg.WriteString("{}")
+	if err != nil {
+		return nil, err
+	}
+
+	return &tc.ContainerRequest{
+		Name:            g.ContainerName,
+		Image:           g.GetImageWithVersion(),
+		Networks:        g.Networks,
+		AlwaysPullImage: true,
+		// ImagePlatform: "linux/x86_64",  //don't even try this on Apple Silicon, the node won't start due to .NET error
+		ExposedPorts: []string{NatPortFormat(DEFAULT_EVM_NODE_HTTP_PORT), NatPortFormat(DEFAULT_EVM_NODE_WS_PORT), NatPortFormat(ETH2_EXECUTION_PORT)},
+		WaitingFor: tcwait.ForAll(
+			tcwait.ForLog("Nethermind initialization completed").
+				WithStartupTimeout(120 * time.Second).
+				WithPollInterval(1 * time.Second),
+		),
+		Cmd: []string{
+			"--datadir=/nethermind",
+			"--config=/none.cfg",
+			fmt.Sprintf("--Init.ChainSpecPath=%s/chainspec.json", GENERATED_DATA_DIR_INSIDE_CONTAINER),
+			"--Init.DiscoveryEnabled=false",
+			"--Init.WebSocketsEnabled=true",
+			fmt.Sprintf("--JsonRpc.WebSocketsPort=%s", DEFAULT_EVM_NODE_WS_PORT),
+			"--JsonRpc.Enabled=true",
+			"--JsonRpc.EnabledModules=net,eth,consensus,subscribe,web3,admin",
+			"--JsonRpc.Host=0.0.0.0",
+			fmt.Sprintf("--JsonRpc.Port=%s", DEFAULT_EVM_NODE_HTTP_PORT),
+			"--JsonRpc.EngineHost=0.0.0.0",
+			"--JsonRpc.EnginePort=" + ETH2_EXECUTION_PORT,
+			fmt.Sprintf("--JsonRpc.JwtSecretFile=%s", JWT_SECRET_FILE_LOCATION_INSIDE_CONTAINER),
+			fmt.Sprintf("--KeyStore.KeyStoreDirectory=%s", KEYSTORE_DIR_LOCATION_INSIDE_CONTAINER),
+			"--KeyStore.BlockAuthorAccount=0x123463a4b065722e99115d6c222f267d9cabb524",
+			"--KeyStore.UnlockAccounts=0x123463a4b065722e99115d6c222f267d9cabb524",
+			fmt.Sprintf("--KeyStore.PasswordFiles=%s", ACCOUNT_PASSWORD_FILE_INSIDE_CONTAINER),
+			"--Network.MaxActivePeers=0",
+			"--Network.OnlyStaticPeers=true",
+			"--HealthChecks.Enabled=true", // default slug /health
+		},
+		Files: []tc.ContainerFile{
+			{
+				HostFilePath:      noneCfg.Name(),
+				ContainerFilePath: "/none.cfg",
+				FileMode:          0644,
+			},
+		},
+		HostConfigModifier: func(hostConfig *container.HostConfig) {
+			hostConfig.Mounts = append(hostConfig.Mounts, mount.Mount{
+				Type:     mount.TypeBind,
+				Source:   g.generatedDataHostDir,
+				Target:   GENERATED_DATA_DIR_INSIDE_CONTAINER,
+				ReadOnly: false,
+			})
+		},
+		LifecycleHooks: []tc.ContainerLifecycleHooks{
+			{
+				PostStarts: g.PostStartsHooks,
+				PostStops:  g.PostStopsHooks,
+			},
+		},
+	}, nil
 }
